Preserve empty source file contents correctly when generating inline source maps
When a source file was added with empty content, the sourceContents
entry in the generated JSON before base64 encoding
was null rather than an empty string.

This causes consumers to try to retrieve the source from the original URL
rather than using the inlined empty content.

In Firefox the presence of a single 'null' entry in the sourceContents
array has additional consequences for how sourcemaps are handled - see
https://github.com/mozilla/source-map/pull/190 for more details.